feat: Organisation endpoints

+++ routes/api.php
@@ -31,80 +31,85 @@ use Illuminate\Support\Facades\Auth;
 
 //Role Admin
 Route::middleware(['auth:api','admin'])->group(function(){
-//Ajouter étudiant cas social
-Route::post('/ajoutEtudiant/CasSocial', '\App\Http\Controllers\UserController@ajoutEtudiantCasSocial');
-//Ajouter un profil utilisateur
-Route::post('/AjouterUtilisateur', '\App\Http\Controllers\UserController@ajoutProfil');
+
+/***Gestion des Utilisateurs */
 //Ajouter un role
 Route::post('/AjouterRole', '\App\Http\Controllers\UserController@ajoutRole');
-
+//Ajouter un profil utilisateur
+Route::post('/AjouterUtilisateur', '\App\Http\Controllers\UserController@ajoutProfil');
 //Modifier l'status d' un profil utilisateur
 Route::put('/modifierProfil/{id}', '\App\Http\Controllers\UserController@ModifierProfil');
+//Lister les profils
+Route::get('/Utilisateurs','\App\Http\Controllers\UserController@listesProfils');
+//Lister un profil chef de pavillon
+Route::get('/Utilisateurs/ChefPavillon/{id}','\App\Http\Controllers\UserController@detailProfilUtilisateurPavillon');
+//Lister un profil chef pedagogique
+Route::get('/Utilisateurs/ChefPedagogique/{id}','\App\Http\Controllers\UserController@detailProfilUtilisateurPedagogique');
+/**Gestion des étudiants cas social */
+//Ajouter étudiant cas social
+Route::post('/ajoutEtudiant/CasSocial', '\App\Http\Controllers\UserController@ajoutEtudiantCasSocial');
+//Liste les étudiants par cas social
+Route::get('/listesEtudiantsCasSocial','\App\Http\Controllers\UserController@listesEtudiantsCasSocial');
+//Lister  detail d'un etudiant
+Route::get('/detailEtudiant/CasSocial/{id}','\App\Http\Controllers\UserController@detailEtudiantCasSocial');
+/* Gestion des etudiants mérites */
+//Lister les etudiants par mérites
+Route::get('/listesEtudiantsMerites','\App\Http\Controllers\UserController@listesEtudiantsMerites');
+//Lister detail d'un étudiant
+Route::get('/detailEtudiant/Merite/{id}','\App\Http\Controllers\UserController@detailEtudiantMerite');
+/* Gestion des pavillons */
 //Lister les Pavillon
 Route::get('/pavillons', '\App\Http\Controllers\PavillonController@index');
-
 //Ajouter un pavillon
 Route::post('/pavillon/create', '\App\Http\Controllers\PavillonController@store');
-
 //Modifier un pavillon
 Route::put('/pavillon/update/{id}', '\App\Http\Controllers\PavillonController@update');
 //Detail un pavillon
 Route::get('/pavillon/read/{id}', '\App\Http\Controllers\PavillonController@show');
 //Supprimer un pavillon
 Route::delete('/pavillon/delete/{id}', '\App\Http\Controllers\PavillonController@destroy');
- //Lister les Chambre
-Route::get('/ListesChambres', '\App\Http\Controllers\ChambreController@index');
-//Lister les etudiants par mérites
-Route::get('/listesEtudiantsMerites','\App\Http\Controllers\UserController@listesEtudiantsMerites');
-//Liste les étudiants par cas social
-Route::get('/listesEtudiantsCasSocial','\App\Http\Controllers\UserController@listesEtudiantsCasSocial');
-//Lister les étudiants par mérite
-Route::get('/listesEtudiantsCasSocial','\App\Http\Controllers\UserController@listesEtudiantsCasSocial');
-//Liste  detail d'un etudiant
-Route::get('/detailEtudiant/CasSocial/{id}','\App\Http\Controllers\UserController@detailEtudiantCasSocial');
-//Lister detail d'un étudiant
-Route::get('/detailEtudiant/Merite/{id}','\App\Http\Controllers\UserController@detailEtudiantMerite');
-//Lister les profils
-Route::get('/Utilisateurs','\App\Http\Controllers\UserController@listesProfils');
-//Lister un profil chef de pavillon
-Route::get('/Utilisateurs/ChefPavillon/{id}','\App\Http\Controllers\UserController@detailProfilUtilisateurPavillon');
+/* Gestion des chambres */
+//Lister les Chambre
+Route::get('/admin/chambres', '\App\Http\Controllers\ChambreController@index');
 //Ajouter une chambre
-Route::post('/chambre/create', '\App\Http\Controllers\ChambreController@store');
+Route::post('/admin/chambre/create', '\App\Http\Controllers\ChambreController@store');
 //Modifier une chambre
-Route::put('/chambre/update/{id}', '\App\Http\Controllers\ChambreController@update');
+Route::put('/admin/chambre/update/{id}', '\App\Http\Controllers\ChambreController@update');
 //Detail chambre
-Route::get('/chambre/read/{id}', '\App\Http\Controllers\ChambreController@show');
+Route::get('/admin/chambre/read/{id}', '\App\Http\Controllers\ChambreController@show');
 //Supprimer une chambre
-Route::delete('/chambre/delete/{id}', '\App\Http\Controllers\ChambreController@destroy');
-
-
-//Lister un profil chef pedagogique
-Route::get('/Utilisateurs/ChefPedagogique/{id}','\App\Http\Controllers\UserController@detailProfilUtilisateurPedagogique');
-
+Route::delete('/admin/chambre/delete/{id}', '\App\Http\Controllers\ChambreController@destroy');
+/**Gestion de validation des codifications */
 //Valider un étudiant non attribuer
 Route::put('/ValiderEtudiant/update/{id}','\App\Http\Controllers\UserController@validerEtudiant');
 });
+/**
+ * Perspectives :
+ *Modifier un étudiant par mérite
+ *Route::put('/ModifierEtudiantsMerites', '\App\Http\Controllers\UserController@modifierEtudiantMerite');
+ *Modifier un étudiant cas social
+ *Route::put('/ModifierEtudiantsCasSocial', '\App\Http\Controllers\UserController@modifierEtudiantCasSocial');
+*/
 //Role = Chef de Service Pédagogique
 Route::middleware(['auth:api','role'])->group(function(){
+/**Gestion des étudiants par mérite */
 //Ajouter un étudiant
 Route::post('/ajoutEtudiant/Merite', '\App\Http\Controllers\UserController@ajoutEtudiantMerite');
-//Modifier un étudiant
-Route::put('/ModifierEtudiantsMerites', '\App\Http\Controllers\UserController@modifierEtudiantMerite');
 // //Lister les étudiants par mérite
 Route::get('/listesEtudiantsMerites','\App\Http\Controllers\UserController@listesEtudiantsMerites');
 //Lister detail d'un etudiant
 Route::get('/detailEtudiant/Merite/{id}','\App\Http\Controllers\UserController@detailEtudiantMerite');
-//Validation mail
-Route::get('/Validation/{email}','\App\Http\Controllers\UserController@SendMailValidation');
 });
 //Role = Chef de pavillon
 Route::middleware(['auth:api','profil'])->group(function(){
+/**Gestion des chambres */
 //Lister les Chambre
 Route::get('/chambres', '\App\Http\Controllers\ChambreController@index');
 //Modifier une chambre
-Route::put('/modifierChambre/update/{id}', '\App\Http\Controllers\ChambreController@update');
+Route::put('/chambre/update/{id}', '\App\Http\Controllers\ChambreController@update');
 //Detail chambre
-Route::get('/detailChambre/read/{id}', '\App\Http\Controllers\ChambreController@show');
+Route::get('/chambre/{id}', '\App\Http\Controllers\ChambreController@show');
+/**Gestion des réclamations des étudiants */
 //Lister les reclamation(s)
 Route::get('/listerDesReclamations', '\App\Http\Controllers\ReclamationController@index');
 //Detail une reclamation
@@ -122,16 +127,16 @@ Route::delete('/SupprimerReclamation/delete/{id}', '\App\Http\Controllers\Reclam
 
 Route::group(['middleware' => 'api', 'prefix' => 'auth'], function ($router) {
 //Login User
-Route::post('login', '\App\Http\Controllers\AuthController@login');
+Route::post('/login', '\App\Http\Controllers\AuthController@login');
 //Logout User
-Route::post('logout', '\App\Http\Controllers\AuthController@logout');
+Route::post('/logout', '\App\Http\Controllers\AuthController@logout');
 //Rafraichir Utilisateur
-Route::post('refresh', '\App\Http\Controllers\AuthController@refresh');
+Route::post('/refresh', '\App\Http\Controllers\AuthController@refresh');
 //Utilisateur Connecter
-Route::post('me', '\App\Http\Controllers\AuthController@me');
+Route::post('/me', '\App\Http\Controllers\AuthController@me');
 });
 //Utilisateur connecté
-Route::get('me', '\App\Http\Controllers\AuthController@me');
+Route::get('/me', '\App\Http\Controllers\AuthController@me');
